Added all project files
mainSnake.py - The base file from which the program is run. 

snakeFunctions.py - Contains a list of functions such as those to update the position of the snake and food use to run the basic snake program

geneticAlgoFunctions.py - Contains a list of functions such as those to get the fitness values for a set of neural nets beings tested or creating a new population of neural nets from a set of parents

initialModel.h5 - A feedforward neural net that has been trained on data obtained from controlling how a snake moves thorugh a handwritten algorithm. Feel free to use your own models -- just make sure that they fit the parameters of those used by the program.

ChoiceData.csv, MovementData.csv - the actual data that was used to train the neural net in the initial model. Choice data has a log of every move made by the snake and movement data has a log of relevant data for the neural net at each position for the snake as it plays a game.